refactor(cli): extract root command composition
